+++ CHANGES.rst
@@ -45,6 +45,9 @@ Enhancements and Fixes
 
 - Tables returned by RegistryResource.get_tables() now have a utype
   attribute [#576]
+ 
+- MIVOT module: If the MIVOT annotation block contains a valid instance of the ``mango:EpochPosition` class,
+  the dynamic object describing the mapped data can generate a valid SkyCoord instance. [#591]
 
 - The MIVOT model is able to automatically build SkyCoord instance from ``mango:EpochPosition`` instances [#591]